pyproject.toml
[project]
name = "core-platform"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.4",
    "ccxt>=4.4.99",
    "celery>=5.5.3",
    "dependency-injector>=4.48.1",
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "prometheus-client>=0.22.1",
    "psutil>=7.0.0",
    "psycopg[binary]>=3.2.9",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "redis>=6.4.0",
    "sqlalchemy>=2.0.42",
    "structlog>=25.4.0",
    "tenacity>=9.1.2",
    "uvicorn[standard]>=0.35.0",
]

[project.scripts]
atlas-api = "app.main:run"

[dependency-groups]
dev = [
    "anyio>=4.10.0",
    "black>=25.1.0",
    "isort>=6.0.1",
    "mypy>=1.17.1",
    "pre-commit>=4.2.0",
    "pytest>=8.4.1",
    "ruff>=0.12.8",
    "types-psutil>=7.0.0.20250801",
    "types-redis>=4.6.0.20241004",
    "types-requests>=2.32.4.20250809",
]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.isort]
profile = "black"
line_length = 100
combine_as_imports = true

[tool.ruff]
line-length = 100
target-version = "py312"
lint.select = [
  "E", "F", "I",  # pycodestyle, pyflakes, isort
  "B",              # flake8-bugbear
  "UP",             # pyupgrade
  "ASYNC",          # flake8-async
]
lint.ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
no_implicit_optional = true
strict_optional = true
check_untyped_defs = true
show_error_codes = true
pretty = true
plugins = []
exclude = ["^alembic/", "^migrations/"]
